requirements.txt
openai>=1.0.0
azure-identity>=1.12.0
python-dotenv>=1.0.0
requests>=2.28.0
jira>=3.4.0
python-docx>=0.8.11
PyPDF2>=3.0.1
beautifulsoup4>=4.11.0
markdown>=3.4.0
pydantic>=2.0.0
rich>=13.0.0
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0
typing-extensions>=4.5.0
openai-agents>=1.0.0 